+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export modal now supports session-only additional notes. Added notes can be included as single notes or as extra roots that run normal traversal from that note.
+
 ## [1.13.0] - 2026-04-29
 
 ### Added